feat(#752):remove redundant code
