+++ src/main/java/org/eolang/jeo/representation/bytecode/BytecodeMethod.java
@@ -439,14 +439,6 @@ public final class BytecodeMethod implements Testable {
      * @return Max local variables.
      */
     private int computeLocals() {
-//        return new MaxLocals(
-//            this.properties,
-//            this.instructions,
-//            this.tryblocks.stream()
-//                .filter(BytecodeTryCatchBlock.class::isInstance)
-//                .map(BytecodeTryCatchBlock.class::cast)
-//                .collect(Collectors.toList())
-//        ).value();
         return new MaxLocals(
             this.properties,
             this.instructions,
